Refactor delisting unit tests to be grouped in only one class
Make it easier to maintain, grow and test
